backend/requirements.txt
# FastAPI y dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
argon2-cffi==23.1.0  # Para hashing seguro de religión

# Validación
pydantic==2.5.0
email-validator==2.1.0.post1

# Configuración
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging y monitoreo
structlog==23.2.0
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.19.0
psutil==5.9.6

# Seguridad adicional
slowapi==0.1.9
cryptography==41.0.7

# Utilidades
validators==0.22.0
python-dateutil==2.8.2
requests==2.31.0
faker==20.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
httpx==0.25.2
coverage==7.3.2
factory-boy==3.3.0

# Desarrollo
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
